+++ deepsig/helpers.py
